refactor(engine): move plugins to their folders
