+++ engine/include/cubos/engine/tools/settings_inspector/plugin.hpp
